Move `CargoCommandLine` back to toolchain package
