+++ src/main/kotlin/org/rust/cargo/project/toolwindow/CargoProjectsTree.kt
@@ -7,10 +7,10 @@ package org.rust.cargo.project.toolwindow
 
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.ui.treeStructure.SimpleTree
-import org.rust.cargo.CargoCommandLine
 import org.rust.cargo.project.model.CargoProject
 import org.rust.cargo.project.toolwindow.CargoProjectTreeStructure.CargoSimpleNode
 import org.rust.cargo.project.workspace.CargoWorkspace
+import org.rust.cargo.toolchain.CargoCommandLine
 import java.awt.event.MouseAdapter
 import java.awt.event.MouseEvent
 import javax.swing.tree.DefaultMutableTreeNode
