+++ src/test/kotlin/org/rust/cargo/runconfig/CargoTestRunStateTest.kt
@@ -11,7 +11,7 @@ import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
-import org.rust.cargo.CargoCommandLine
+import org.rust.cargo.toolchain.CargoCommandLine
 import org.rust.cargo.toolchain.RustChannel
 import org.rust.cargo.toolchain.impl.RustcVersion
 import java.nio.file.Path
