+++ src/test/kotlin/org/rust/cargo/toolchain/CargoCommandLineTest.kt
@@ -7,7 +7,6 @@ package org.rust.cargo.toolchain
 
 import org.junit.Assert
 import org.junit.Test
-import org.rust.cargo.CargoCommandLine
 import java.nio.file.Paths
 
 class CargoCommandLineTest {
