+++ src/test/kotlin/org/rust/cargo/toolchain/CargoTest.kt
@@ -10,7 +10,6 @@ import com.intellij.openapi.util.SystemInfo
 import com.intellij.util.net.HttpConfigurable
 import org.rust.MockRustcVersion
 import org.rust.RsTestBase
-import org.rust.cargo.CargoCommandLine
 import org.rust.cargo.project.settings.rustSettings
 import org.rust.cargo.toolchain.tools.cargo
 import java.nio.file.Paths
